+++ src/oandapysuite/objects/trade.py
@@ -192,25 +192,25 @@ class Backtester(MarketSimulator):
 
     def __check_signal(self):
         cands = self.get_candle_cluster()
-        mvdiff_obj = self.signal()
-        mvdiff = mvdiff_obj.get_signal(self.current_candle, [self.altavd, self.sma])
-        print(mvdiff)
+        mvdiff = self.sig_mvdiff.get_signal(self.current_candle, [self.altavd, self.sma])
         if self.trade_type == 0:
             if mvdiff == 1:
-                print(f'Entering long @ {self.current_price}, time: {self.current_time}')
+                print(f'Entering long @ {self.current_price}, time: {self.current_time} TP:{self.sig_mvdiff.take_profit}, SL:{self.sig_mvdiff.stop_loss}')
+                self.entry_price = Decimal(self.current_price)
                 self.__enter_trade(trade_type=1)
-            if mvdiff == 2:
-                print(f'Entering short @ {self.current_price}, time: {self.current_time}')
-                self.__enter_trade(trade_type=2)
+            if mvdiff == 3:
+                print(f'Entering short @ {self.current_price}, time: {self.current_time}, TP:{self.sig_mvdiff.take_profit}, SL:{self.sig_mvdiff.stop_loss}')
+                self.entry_price = Decimal(self.current_price)
+                self.__enter_trade(trade_type=3)
         if self.trade_type != 0:
-            if self.trade_type == 1 and mvdiff == 3:
+            if self.trade_type == 1 and mvdiff == 2:
                 self.__exit_trade()
-                print(f'Exiting long, profit:{(self.current_price - self.entry_price) * 10000}  time: {self.current_time}')
-                self.trades.append((self.current_price - self.entry_price) * 10000)
-            if self.trade_type == 2 and mvdiff == 4:
+                print(f'Exiting long, price:{self.current_price}  time: {self.current_time}')
+                self.trades.append((self.current_price - self.entry_price))
+            if self.trade_type == 3 and mvdiff == 4:
                 self.__exit_trade()
-                print(f'Exiting long, profit:{(self.entry_price - self.current_price) * 10000}  time: {self.current_time}')
-                self.trades.append((self.entry_price - self.current_price) * 10000)
+                print(f'Exiting short, price:{self.current_price}  time: {self.current_time}')
+                self.trades.append((self.entry_price - self.current_price))
 
     def __update_price(self, candle, is_close=False, is_open=False):
         if is_close: self.current_price = candle.close
@@ -223,7 +223,7 @@ class Backtester(MarketSimulator):
             self.current_price += candle.low + up_or_down() *((candle.high-candle.low) * Decimal(ceil(random()*20)*5/200))
             if self.current_price > candle.high: self.current_price = candle.high
             elif self.current_price < candle.low: self.current_price = candle.low
-        if self.periods > 201:
+        if self.periods > 30:
             self.__check_signal()
 
     def __get_candle_at_time(self, candle, target_time):
@@ -265,8 +265,9 @@ class Backtester(MarketSimulator):
         self.periods = 0
 
     def run(self):
-        self.altavd = self.indicators[1](on='open', period=100, name='altav', color='green')
-        self.sma = self.indicators[0](on='open', period=200, name='sma', color='black')
+        self.altavd = self.indicators[1](on='open', period=9, name='altav', color='green')
+        self.sma = self.indicators[0](on='open', period=24, name='sma', color='black')
+        self.sig_mvdiff = self.signal()
         for i in range(len(self.window)):
             current_macro_candle = self.window[i]
             while self.current_time < current_macro_candle.time + timedelta(seconds=candlex[current_macro_candle.gran]):